requirements.txt
Django==2.1.5
gunicorn==19.6.0  
psycopg2==2.7.5
celery==4.2.1
redis==2.10.6
django-taggit
django-social-share
social-auth-app-django==2.1.0
markdown
pillow
sorl-thumbnail==12.4.1
django-kronos

